Agedetect.java
Write a program that asks a user for their birth year encoded as two digits (like "62") and for the current year, also encoded as two digits (like "99"). The program is to correctly write out the user’s age in years.

Input Format:

Input consists of 2 integers. 

The first integer corresponds to the last 2 digits of the birth year. 

The second integer corresponds to the last 2 digits of the current year.

Output Format:

Refer sample input and output for further formatting specifications.

[All text in bold corresponds to input and the rest corresponds to output]

Sample Input and Output:

Enter Year of Birth

62

Enter Current year

99

Your age is 37