count letters and digits in a string
Write a program that accepts a sentence and calculate the number of letters and digits.

Suppose the following input is supplied to the program:

Input Format:

Input consists of a string. 

Output Format: 

The output consists of a calculation of number of digits and letters. 

Sample Input: 

hello world! 123

Sample output:

LETTERS 10

DIGITS 3