got furniture and sell
Depreciation is a method of reallocating the cost of a tangible asset over its useful lifespan of it being in motion. Businesses depreciate long-term assets for both tax and accounting purposes.
Input format:
First input corresponds to the principle of the furniture.
Second  input represents the percentage of depreciation. 
Third input represents number of years

Sample Input: 
10000
5
5
Sample Output:
7736